test(formatters): relocate mocked formatter provider
