+++ packages/suite/src/support/tests/hooksHelper.tsx
@@ -13,7 +13,7 @@ import {
 import userEvent from '@testing-library/user-event';
 
 import { ServicesProvider } from '@suite-common/dependency-injection';
-import { MockedFormatterProvider } from '@suite-common/formatters';
+import { MockedFormatterProvider } from '@suite-common/formatters/mocks';
 
 import { ConnectedThemeProvider } from 'src/support/suite/ConnectedThemeProvider';
