+++ suite-common/formatters/mocks/index.ts
@@ -1,0 +1,1 @@
+export { MockedFormatterProvider } from './MockedFormatterProvider';
