+++ suite-common/formatters/package.json
@@ -5,6 +5,16 @@
     "license": "See LICENSE.md in repo root",
     "sideEffects": false,
     "main": "src/index",
+    "exports": {
+        ".": {
+            "types": "./libDev/src/index.d.ts",
+            "default": "./src/index.ts"
+        },
+        "./mocks": {
+            "types": "./libDev/mocks/index.d.ts",
+            "default": "./mocks/index.ts"
+        }
+    },
     "scripts": {
         "test:unit": "yarn g:jest",
         "depcheck": "yarn g:depcheck",
