+++ suite/test-utils/src/BasicProviderForTests.tsx
@@ -2,7 +2,7 @@ import type { PropsWithChildren } from 'react';
 
 import { IntlProviderForTests } from '@suite/intl';
 import { ServicesProvider } from '@suite-common/dependency-injection';
-import { MockedFormatterProvider } from '@suite-common/formatters';
+import { MockedFormatterProvider } from '@suite-common/formatters/mocks';
 import { ConnectedThemeProvider, ResponsiveContextProvider } from '@trezor/suite';
 import { extraDependenciesDesktopMock } from '@trezor/suite/mocks';
